+++ src/Plugin.php
@@ -30,7 +30,7 @@ defined('ABSPATH') || exit;
 
 final class Plugin
 {
-    public const VERSION = '0.1.2';
+    public const VERSION = '0.1.3';
     public const PLUGIN_ID = 'uniple';
 
     public static function boot(): void